ESC-710 initial EmailVerificationSupport trait
